Dynarec: Fixed memory manager
Not all references to compiled blocks were removed when cleaning
segments because it was using the size of a block instead of the size of
a segment.

This was causing a core crash after the cache was full as the dynarec
was trying to jump in a compiled block that has been overwritten by a
new compiled block.

Not all games do actually reach this cache limit, but zelda games do.
That's why zelda games were constantly crashing after ~30min.

From new_dynarec.txt

==Memory manager==

The last step in new_recompile_block is to ensure that there will be
sufficient memory available to compile the next block.  If there is not,
then the oldest blocks are purged.

The dynarec cache can be described as a 32MB circular buffer divided
into eight segments.  Memory is allocated in order from beginning to
end.

When there are less than 2 empty segments, the next segment in sequence
is cleared, wrapping around to the beginning of the buffer.  This
continues as memory is needed, wrapping around from end to beginning.

(cherry picked from commit 60a12ce73ab644ce56895c116763a33806214144)